1.0.9

+++ cloud_storage_pwc/AzureStorage/DataLake/__init__.py
@@ -155,15 +155,15 @@ class DataLake(StorageAccountVirtualClass):
             main_directory = file_system_client.get_directory_client("/")
             if directory_path != "":
                 main_directory = main_directory.get_sub_directory_client(directory_path)
-
-            main_directory.delete_directory()
+            if main_directory.exists():
+                main_directory.delete_directory()
             
-            if wait :
-                time.sleep(1)
-                dict_exists = main_directory.exists()
-                while dict_exists:
-                    time.sleep(2)
+                if wait :
+                    time.sleep(1)
                     dict_exists = main_directory.exists()
+                    while dict_exists:
+                        time.sleep(2)
+                        dict_exists = main_directory.exists()
         except ContainerNotFound as e:
             raise e
         except Exception as e:
